fix: Re-running flows tutorial with pre-existing ACL
This is mostly a change for maintainers. It's easy to forget to
delete the ACL at the end of the notebook, which will cause the
original flow to fail due to the ACL alerady existing. If the ACL
isn't in the flow output, fetch it using the destination ACL path.

Flow exception code to handle this is left out on purpose to make
the flow simpler to understand.